update finally working properly
